🐛fix/detail 폴더 main으로 이름 변경 및 컴포넌트 분리 작업
