+++ src/main/DetailStationExit.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+function DetailStationExit() {
+  return (
+    <div className="exit-information">
+      {/* <span className="exit-title">5번 출구</span> */}
+      {/* <ul className="exit-list">
+      <li>
+        <span>홍익대학교 서울 캠퍼스</span>
+        <span>10분</span>
+      </li>
+      <li>
+        <span>홍익대학교 서울 캠퍼스</span>
+        <span>10분</span>
+      </li>
+    </ul> */}
+    </div>
+  );
+}
+
+export default DetailStationExit;
