+++ src/main/Main.tsx
@@ -1,12 +1,12 @@
 import { TransformWrapper, TransformComponent } from 'react-zoom-pan-pinch';
 // components
 import '../styles/detail.scss';
-import DetailStationResult from './DetailStationResult';
+import DetailStationResult from './DetailStation';
 import Line2 from '../map/Line2';
 import Line4 from '../map/Line4';
 import Line5 from '../map/Line5';
 
-function DetailStation() {
+function Main() {
   return (
     <div className="detail-station-wrap">
       <div className="detail-station-contatiner">
@@ -35,4 +35,4 @@ function DetailStation() {
   );
 }
 
-export default DetailStation;
+export default Main;
